type error fixed

+++ CookieData/Migrations/20230115212321_InitialCreate.cs
@@ -19,7 +19,7 @@ namespace CookieData.Migrations
                     GameAccountId = table.Column<int>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Networth = table.Column<double>(type: "float", nullable: false),
-                    Cookies = table.Column<long>(type: "bigint", nullable: false),
+                    Cookies = table.Column<double>(type: "float", nullable: false),
                     Clicks = table.Column<long>(type: "bigint", nullable: false)
                 },
                 constraints: table =>
